Fix: Use one GetHostAddr only: LibGetHostAddr in mdsshr
